02 - LLM API with Python, token streaming/requirements.txt
openai==1.57.3
python-dotenv==1.0.1
